+++ applications/fishing-map/src/features/workspace/vessels/VesselLayerPanel.tsx
@@ -28,22 +28,6 @@ import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
 import ExpandedContainer from 'features/workspace/shared/ExpandedContainer'
 import { Bbox } from 'types'
 
-// Translations by feature.unit static keys
-// t('vessel.flag', 'Flag')
-// t('vessel.imo', 'IMO')
-// t('vessel.firstTransmissionDate', 'First transmission date')
-// t('vessel.lastTransmissionDate', 'Last transmission date')
-// t('vessel.registeredGearType', 'Registered Gear Type')
-// t('vessel.widthRange', 'Width range')
-// t('vessel.lengthRange', 'Length range')
-// t('vessel.grossTonnageRange', 'Gross Tonnage range')
-// t('vessel.fleet', 'Fleet')
-// t('vessel.source', 'Source')
-// t('vessel.nationalId', 'National Id')
-// t('vessel.length', 'Length')
-// t('vessel.beam', 'Beam')
-// t('vessel.capacity', 'Capacity')
-
 type LayerPanelProps = {
   dataview: UrlDataviewInstance
 }
@@ -130,6 +114,31 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
   const infoError = infoResource?.status === ResourceStatus.Error
   const trackError = trackResource?.status === ResourceStatus.Error
 
+  const getFieldValue = (field: any, fieldValue: string) => {
+    if (field.type === 'date') {
+      return <I18nDate date={fieldValue} />
+    }
+    if (field.type === 'flag') {
+      return <I18nFlag iso={fieldValue} />
+    }
+    if (field.id === 'geartype') {
+      return t(`vessel.gearTypes.${fieldValue}` as any, '---')
+    }
+    if (field.id === 'mmsi') {
+      return (
+        <a
+          className={styles.link}
+          target="_blank"
+          rel="noreferrer"
+          href={`https://www.marinetraffic.com/en/ais/details/ships/${fieldValue}`}
+        >
+          {formatInfoField(fieldValue, field.type)}
+        </a>
+      )
+    }
+    return formatInfoField(fieldValue, field.type)
+  }
+
   return (
     <div
       className={cx(styles.LayerPanel, { [styles.expandedContainerOpen]: colorOpen || infoOpen })}
@@ -210,22 +219,7 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
                           <label>{t(`vessel.${field.id}` as any)}</label>
                           {fieldValues.map((fieldValue, i) => (
                             <span key={fieldValue}>
-                              {field.type === 'date' ? (
-                                <I18nDate date={fieldValue} />
-                              ) : field.type === 'flag' ? (
-                                <I18nFlag iso={fieldValue} />
-                              ) : field.id === 'mmsi' ? (
-                                <a
-                                  className={styles.link}
-                                  target="_blank"
-                                  rel="noreferrer"
-                                  href={`https://www.marinetraffic.com/en/ais/details/ships/${fieldValue}`}
-                                >
-                                  {formatInfoField(fieldValue, field.type)}
-                                </a>
-                              ) : (
-                                formatInfoField(fieldValue, field.type)
-                              )}
+                              {getFieldValue(field, fieldValue)}
                               {/* Field values separator */}
                               {i < fieldValues.length - 1 ? ', ' : ''}
                             </span>