+++ packages/fourwings-aggregate/src/index.ts
@@ -8,3 +8,11 @@ export {
   getCellArrayIndex,
   generateUniqueId,
 } from './util'
+export {
+  GeomType,
+  SublayerCombinationMode,
+  AggregationOperation,
+  TileAggregationParams,
+  TileAggregationSourceParams,
+} from './types'
+export { VALUE_MULTIPLIER } from './constants'