+++ packages/timebar/types/immediateContext.d.ts
@@ -1,0 +1,2 @@
+export default ImmediateContext;
+declare const ImmediateContext: any;