+++ packages/timebar/types/utils/index.d.ts
@@ -1,0 +1,12 @@
+export function getHumanizedDates(start: any, end: any): {
+    humanizedStart: any;
+    humanizedEnd: any;
+    interval: any;
+};
+export function getTimebarRangeByWorkspace(timeline: any): {
+    start: any;
+    end: any;
+};
+export function geoJSONTrackToTimebarFeatureSegments({ features }?: {
+    features?: any[];
+}): any[][];